+++ tests/python/rest/testcases/test_add_simple.py
@@ -1,0 +1,10 @@
+import unittest2 as unittest
+
+from client import make_rest_call
+
+
+class AddSimpleTestCase(unittest.TestCase):
+
+    def test_simple(self):
+        self.assertEquals(
+            make_rest_call('add_simple', a='hello', b=' world'), 'hello world')